[Web3j] Implementing Web3J datasource layer

+++ app/src/main/java/com/dreamsoftware/artcollectibles/data/blockchain/entity/ArtCollectibleBlockchainEntity.kt
@@ -1,7 +1,9 @@
 package com.dreamsoftware.artcollectibles.data.blockchain.entity
 
+import java.math.BigInteger
+
 data class ArtCollectibleBlockchainEntity(
     val creator: String,
-    val royalty: Int,
+    val royalty: BigInteger,
     val isExist: Boolean
 )
